requirements.txt
Mezzanine==3.1.10
django-contact-form==1.0
